Use Dependabot to check that the base image is up to date
Extends #18 with Dependabot checking the base image to make sure it is up to date weekly.
Github actions checks each push to make sure the image can be build, launched, and will return a 200 from the default ERDDAP page.
When a commit is to the main or master branch, then the built image will be pushed to Docker Hub (if appropriate secrets are setup, and it is to Axiom's repo).