Avoid duplicate CI runs
Since we combined the release workflow with the push workflow we lost
the ability to use the bracket style "run on just one of these". This
will now only run CI for pushes to the `main` branch, PRs, and releases.

Also ordered the tags differently so the CI logs show which docker image
is being tested and not just "latest"

Removed references to "master"